Renamed CxbxVSBC -> vsbc
