+++ src/vsbc/dllmain.cpp
