+++ src/vsbc/stdafx.cpp
