+++ src/vsbc/stdafx.h
